+++ addons/account/models/company.py
@@ -1,15 +1,12 @@
 # -*- coding: utf-8 -*-
 
 from datetime import timedelta, datetime
-from dateutil.relativedelta import relativedelta
-
 import calendar
-import time
 
 from odoo import fields, models, api, _
-from odoo.exceptions import ValidationError, UserError
+from odoo.exceptions import ValidationError
+from odoo.exceptions import UserError
 from odoo.tools.float_utils import float_round, float_is_zero
-from odoo.tools.misc import DEFAULT_SERVER_DATE_FORMAT
 
 
 class ResCompany(models.Model):
@@ -65,40 +62,6 @@ Best Regards,''')
     account_setup_fy_data_done = fields.Boolean('Financial Year Setup Marked As Done', help="Technical field holding the status of the financial year setup step.")
     account_setup_coa_done = fields.Boolean(string='Chart of Account Checked', help="Technical field holding the status of the chart of account setup step.")
     account_setup_bar_closed = fields.Boolean(string='Setup Bar Closed', help="Technical field set to True when setup bar has been closed by the user.")
-
-    @api.constrains('period_lock_date', 'fiscalyear_lock_date')
-    def _check_lock_dates(self):
-        period_lock_date = self.period_lock_date and\
-            time.strptime(self.period_lock_date, DEFAULT_SERVER_DATE_FORMAT) or False
-        old_fiscalyear_lock_date = self._context.get('old_fiscalyear_lock_date') and\
-            time.strptime(self._context['old_fiscalyear_lock_date'], DEFAULT_SERVER_DATE_FORMAT) or False
-        fiscalyear_lock_date = self.fiscalyear_lock_date and\
-            time.strptime(self.fiscalyear_lock_date, DEFAULT_SERVER_DATE_FORMAT) or False
-
-        # Check the irreversibility of the lock date for all users
-        if old_fiscalyear_lock_date:
-            if fiscalyear_lock_date:
-                if fiscalyear_lock_date < old_fiscalyear_lock_date:
-                    raise ValidationError(_('The lock date for all users is irreversible and must be strictly higher than the previous date.'))
-            else:
-                raise ValidationError(_('The lock date for all users is irreversible and can\'t be removed'))
-
-        if not fiscalyear_lock_date:
-            return
-
-        # Check the lock date for all users
-        previous_month = datetime.strptime(fields.Date.today(), DEFAULT_SERVER_DATE_FORMAT) + relativedelta(months=-1)
-        days_previous_month = calendar.monthrange(previous_month.year, previous_month.month)
-        previous_month = previous_month.replace(day=days_previous_month[1]).timetuple()
-        if fiscalyear_lock_date <= previous_month:
-            raise ValidationError(_('The lock date for all users must be higher than the last day of the previous month.'))
-
-        if not period_lock_date:
-            return
-
-        # Check the lock date for non-advisers
-        if period_lock_date > fiscalyear_lock_date:
-            raise ValidationError(_('The lock date for all users must be higher or equal to the lock date for non-advisers.'))
 
     @api.model
     def _verify_fiscalyear_last_day(self, company_id, last_day, last_month):
