+++ addons/account_lock_constrains/__manifest__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+{
+    'name' : 'Irreversible Lock Date',
+    'version' : '1.0',
+    'category': 'Accounting',
+    'depends' : ['account'],
+    'data': [],
+}
